requirements.txt
einops
faster-coco-eval
graphviz
hydra-core
loguru
numpy
opencv-python
Pillow
pycocotools
requests
rich
torch
torchvision
wandb
